# cli/api/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"XReserveTokenInfo": {
				"properties": {
					"bridgeAddress": {
						"type": "string"
					},
					"feeConst": {
						"type": "string"
					},
					"feeShare": {
						"type": "string"
					},
					"protocolAddress": {
						"type": "string"
					}
				},
				"required": [
					"bridgeAddress",
					"feeConst",
					"feeShare"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TransferTime": {
				"$ref": "#/components/schemas/Record_string.MessengerTransferTime_"
			},
...
			"HttpExceptionBodyMessage": {
				"anyOf": [
					{
						"type": "string"
					},
					{
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					{
						"type": "number",
						"format": "double"
					}
				]
			},
			"HttpExceptionBody": {
				"properties": {
					"message": {
						"$ref": "#/components/schemas/HttpExceptionBodyMessage"
					},
					"error": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AmountFormatted": {
				"properties": {
					"int": {
						"type": "string"
					},
					"float": {
						"type": "string"
					}
				},
				"required": [
					"int",
					"float"
				],
				"type": "object",
				"description": "Describes the same amount in two variation formats.\n\nFor {@link AmountFormat.INT} value contains the amount in the smallest denomination\n\nFor {@link AmountFormat.FLOAT} value contains the amount in tokens denomination"
			},
...
			"Record_string.ChainDetailsWithTokens_": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/ChainDetailsWithTokens"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ChainDetailsMap": {
				"$ref": "#/components/schemas/Record_string.ChainDetailsWithTokens_"
			},
			"CYDToken": {
				"properties": {
					"symbol": {
						"type": "string",
						"description": "Token symbol"
					},
					"name": {
						"type": "string",
						"description": "Token name"
					},
					"decimals": {
						"type": "number",
						"format": "double",
						"description": "Token decimals"
					},
					"poolAddress": {
						"type": "string",
						"description": "Token pool address"
					},
					"tokenAddress": {
						"type": "string",
						"description": "Token address"
					},
					"originTokenAddress": {
						"type": "string",
						"description": "Token origin address"
					},
					"cctpAddress": {
						"type": "string",
						"description": "Token CCTP address</br>\nOptional. Defined if CCTP supported by token"
					},
					"cctpFeeShare": {
						"type": "string",
						"description": "Token CCTP V2 address</br>\nOptional. Defined if CCTP supported by token"
					},
					"cctpV2Address": {
						"type": "string",
						"description": "Token CCTP address</br>\nOptional. Defined if CCTP supported by token"
					},
					"cctpV2FeeShare": {
						"type": "string",
						"description": "Token CCTP V2 address</br>\nOptional. Defined if CCTP supported by token"
					},
					"xReserve": {
						"$ref": "#/components/schemas/XReserveTokenInfo",
						"description": "XReserve bridge configuration for token.\nOptional. Defined if XReserve supported by token"
					},
					"oftId": {
						"type": "string",
						"description": "Internal identifier for the same token across different chains.\n\nUsed to link representations of a token deployed on multiple networks (e.g., Ethereum, BSC).\nTokens with the same `oftId` are considered to be equivalent.\n\nOptional. Defined if OFT supported by token"
					},
					"feeShare": {
						"type": "string",
						"description": "Token fee share"
					},
					"apr": {
						"type": "string"
					},
					"apr7d": {
						"type": "string",
						"description": "Token APR based on last 7 days"
					},
					"apr30d": {
						"type": "string",
						"description": "Token APR based on last 30 days"
					},
					"lpRate": {
						"type": "string",
						"description": "Token LP rate"
					},
					"yieldId": {
						"type": "number",
						"format": "double"
					},
					"suiAddresses": {
						"$ref": "#/components/schemas/SuiAddresses",
						"description": "Sui addresses\nOptional. Defined for SUI"
					},
					"allbridgeChainId": {
						"type": "number",
						"format": "double",
						"description": "Allbridge's Id"
					},
					"bridgeId": {
						"type": "string",
						"description": "Algo Bridge appId"
					},
					"paddingUtilId": {
						"type": "string",
						"description": "Algo Padding appId"
					},
					"bridgeAddress": {
						"type": "string",
						"description": "Bridge contract address"
					},
					"oftBridgeAddress": {
						"type": "string",
						"description": "OFT Bridge contract address\nOptional. Defined if the chain supports OFT"
					},
					"yieldAddress": {
						"type": "string",
						"description": "Yield contract address"
					},
					"abrPayer": {
						"$ref": "#/components/schemas/AbrPayerChainInfo",
						"description": "ABR payer chain info"
					},
					"transferTime": {
						"$ref": "#/components/schemas/TransferTime",
						"description": "Average transfer time to other blockchains"
					},
					"txCostAmount": {
						"$ref": "#/components/schemas/TxCostAmount",
						"description": "Transfers costs"
					},
					"confirmations": {
						"type": "number",
						"format": "double",
						"description": "Number of confirmations required"
					},
					"chainSymbol": {
						"$ref": "#/components/schemas/ChainSymbol",
						"description": "The symbol of the chain representing one of the supported blockchain networks (e.g., \"ETH\" for Ethereum). For more details, see: {@link ChainSymbol}."
					},
					"chainId": {
						"type": "string",
						"description": "Common Blockchain Id</br>\nA 0x-prefixed hexadecimal string</br>\nOptional."
					},
					"chainType": {
						"$ref": "#/components/schemas/ChainType",
						"description": "Blockchain type. For more details, see: {@link ChainType}."
					},
					"chainName": {
						"type": "string",
						"description": "Blockchain network name"
					},
					"tokens": {
						"items": {
							"$ref": "#/components/schemas/TokenWithChainDetailsYield"
						},
						"type": "array"
					}
				},
				"required": [
					"symbol",
					"name",
					"decimals",
					"poolAddress",
					"tokenAddress",
					"feeShare",
					"apr",
					"apr7d",
					"apr30d",
					"lpRate",
					"allbridgeChainId",
					"bridgeAddress",
					"transferTime",
					"txCostAmount",
					"confirmations",
					"chainSymbol",
					"chainType",
					"chainName",
					"yieldAddress",
					"tokens"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GasFeeOptions": {
				"description": "Describes available options of paying the gas fee and the amount to pay when using the corresponding method.\n\nFor {@link FeePaymentMethod.WITH_NATIVE_CURRENCY} value contains the amount in the smallest denomination of the source chain currency\n\nFor {@link FeePaymentMethod.WITH_STABLECOIN} value contains the amount in the smallest denomination of the source token",
				"properties": {
					"native": {
						"$ref": "#/components/schemas/AmountFormatted"
					},
					"stablecoin": {
						"$ref": "#/components/schemas/AmountFormatted"
					},
					"abr": {
						"$ref": "#/components/schemas/AmountFormatted"
					},
					"adminFeeShareWithExtras": {
						"type": "string"
					}
				},
				"required": [
					"native"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GasBalanceResponse": {
				"properties": {
					"gasBalance": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/AddressStatus"
					}
				},
				"required": [
					"gasBalance",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ExtraGasMaxLimitResponse": {
				"description": "Provide extra gas information",
				"properties": {
					"extraGasMax": {
						"$ref": "#/components/schemas/ExtraGasMaxLimits",
						"description": "See {@link ExtraGasMaxLimits}"
					},
					"destinationChain": {
						"properties": {
							"transfer": {
								"$ref": "#/components/schemas/AmountFormatted",
								"description": "cost of send tx on chain"
							},
							"swap": {
								"$ref": "#/components/schemas/AmountFormatted",
								"description": "cost of swap tx on chain"
							},
							"gasAmountMax": {
								"$ref": "#/components/schemas/ExtraGasMaxLimit",
								"description": "gasAmountMax maximum amount you can receive as extra gas on dest chain"
							}
						},
						"required": [
							"transfer",
							"swap",
							"gasAmountMax"
						],
						"type": "object",
						"description": "Information due to destination chain"
					},
					"exchangeRate": {
						"type": "string",
						"description": "Exchange rate"
					},
					"abrExchangeRate": {
						"type": "string",
						"description": "Exchange rate"
					},
					"sourceNativeTokenPrice": {
						"type": "string",
						"description": "Source native token price"
					}
				},
				"required": [
					"extraGasMax",
					"destinationChain",
					"exchangeRate",
					"sourceNativeTokenPrice"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "rest-api",
		"version": "3.30.0-alpha.2",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/chains": {
			"get": {
				"operationId": "ChainDetailsMap",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ChainDetailsMap"
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"swap",
								"pool"
							]
						}
					}
				]
			}
		},
		"/tokens": {
			"get": {
				"operationId": "GetTokens",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/CYDToken"
											},
											"type": "array"
										},
										{
											"items": {
												"$ref": "#/components/schemas/TokenWithChainDetails"
											},
											"type": "array"
										}
									]
								},
								"examples": {
									"Supported tokens": {
										"value": [
											{
												"symbol": "USDC",
												"name": "USD Coin",
												"decimals": 6,
												"poolAddress": "0x0000000000000000000000000000000000000002",
												"tokenAddress": "0x0000000000000000000000000000000000000001",
												"feeShare": "0.001",
												"apr": "0.12",
												"apr7d": "0.1234",
												"apr30d": "0.1199",
												"lpRate": "1.01",
												"chainSymbol": "ETH",
												"chainType": "EVM",
												"chainId": "1",
												"chainName": "Ethereum",
												"allbridgeChainId": 1,
												"bridgeAddress": "0x0000000000000000000000000000000000000003",
												"confirmations": 12,
												"txCostAmount": {
													"transfer": "5000",
													"swap": "1000"
												},
												"transferTime": {
													"ALLBRIDGE": 90000
												}
											}
										]
									},
									"Supported tokens filtered by chain": {
										"value": [
											{
												"symbol": "USDC",
												"name": "USD Coin",
												"decimals": 6,
												"poolAddress": "0x0000000000000000000000000000000000000002",
												"tokenAddress": "0x0000000000000000000000000000000000000001",
												"feeShare": "0.001",
												"apr": "0.12",
												"apr7d": "0.1234",
												"apr30d": "0.1199",
												"lpRate": "1.01",
												"chainSymbol": "ETH",
												"chainType": "EVM",
												"chainId": "1",
												"chainName": "Ethereum",
												"allbridgeChainId": 1,
												"bridgeAddress": "0x0000000000000000000000000000000000000003",
												"confirmations": 12,
												"txCostAmount": {
													"transfer": "5000",
													"swap": "1000"
												},
												"transferTime": {
													"ALLBRIDGE": 90000
												}
											}
										]
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"swap",
								"pool",
								"yield"
							]
						}
					},
					{
						"in": "query",
						"name": "chain",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"BSC",
								"ETH",
								"BAS",
								"SOL",
								"TRX",
								"POL",
								"ARB",
								"CEL",
								"AVA",
								"SRB",
								"STLR",
								"OPT",
								"SUI",
								"SNC",
								"UNI",
								"LIN",
								"ALG",
								"STX"
							]
						}
					}
				]
			}
		},
		"/token/balance": {
			"get": {
				"operationId": "GetTokenBalance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"result": {
											"type": "string"
										}
									},
									"required": [
										"result"
									],
									"type": "object"
								},
								"examples": {
									"Token balance in smallest units": {
										"value": {
											"result": "1000000"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/token/native/balance": {
			"get": {
				"operationId": "GetTokenNativeBalance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AmountFormatted"
								},
								"examples": {
									"Native balance response": {
										"value": {
											"int": "1000000",
											"float": "1"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chain",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/token/details": {
			"get": {
				"operationId": "GetTokenByChainAndAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CYDToken"
										},
										{
											"$ref": "#/components/schemas/TokenWithChainDetails"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chain",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"BSC",
								"ETH",
								"BAS",
								"SOL",
								"TRX",
								"POL",
								"ARB",
								"CEL",
								"AVA",
								"SRB",
								"STLR",
								"OPT",
								"SUI",
								"SNC",
								"UNI",
								"LIN",
								"ALG",
								"STX"
							]
						}
					}
				]
			}
		},
		"/gas/fee": {
			"get": {
				"operationId": "GetGasFeeOptions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GasFeeOptions"
								},
								"examples": {
									"Gas fee options": {
										"value": {
											"native": {
												"int": "5000",
												"float": "0.000005"
											},
											"stablecoin": {
												"int": "1200",
												"float": "0.0012"
											},
											"abr": {
												"int": "980",
												"float": "0.00098"
											},
											"adminFeeShareWithExtras": "0.1"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sourceToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "destinationToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "messenger",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"ALLBRIDGE",
								"WORMHOLE",
								"CCTP",
								"CCTP_V2",
								"OFT",
								"X_RESERVE"
							]
						}
					}
				]
			}
		},
		"/gas/balance": {
			"get": {
				"operationId": "GetGasBalance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GasBalanceResponse"
								},
								"examples": {
									"Gas balance response": {
										"value": {
											"gasBalance": "0.125",
											"status": "OK"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "chain",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"BSC",
								"ETH",
								"BAS",
								"SOL",
								"TRX",
								"POL",
								"ARB",
								"CEL",
								"AVA",
								"SRB",
								"STLR",
								"OPT",
								"SUI",
								"SNC",
								"UNI",
								"LIN",
								"ALG",
								"STX"
							]
						}
					}
				]
			}
		},
		"/gas/extra/limits": {
			"get": {
				"operationId": "GetExtraGasMaxLimits",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExtraGasMaxLimitResponse"
								},
								"examples": {
									"Extra gas limit response": {
										"value": {
											"extraGasMax": {
												"native": {
													"int": "1000000",
													"float": "0.001"
												},
												"stablecoin": {
													"int": "500000",
													"float": "0.5"
												},
												"abr": {
													"int": "490000",
													"float": "0.49"
												}
											},
											"destinationChain": {
												"gasAmountMax": {
													"int": "250000",
													"float": "0.00025"
												},
												"swap": {
													"int": "1000",
													"float": "0.000001"
												},
												"transfer": {
													"int": "2000",
													"float": "0.000002"
												}
											},
											"exchangeRate": "1.02",
											"abrExchangeRate": "0.98",
											"sourceNativeTokenPrice": "3200.15"
										}
									},
									"Extra gas limit response for a specific messenger": {
										"value": {
											"extraGasMax": {
												"native": {
													"int": "1000000",
													"float": "0.001"
												},
												"stablecoin": {
													"int": "500000",
													"float": "0.5"
												},
												"abr": {
													"int": "490000",
													"float": "0.49"
												}
											},
											"destinationChain": {
												"gasAmountMax": {
													"int": "250000",
													"float": "0.00025"
												},
												"swap": {
													"int": "1000",
													"float": "0.000001"
												},
												"transfer": {
													"int": "2000",
													"float": "0.000002"
												}
											},
											"exchangeRate": "1.02",
											"abrExchangeRate": "0.98",
											"sourceNativeTokenPrice": "3200.15"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sourceToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "destinationToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "messenger",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"ALLBRIDGE",
								"WORMHOLE",
								"CCTP",
								"CCTP_V2",
								"OFT",
								"X_RESERVE"
							]
						}
					}
				]
			}
		},
		"/check/stellar/balanceline": {
			"get": {
				"operationId": "CheckBalanceLine",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HorizonApi.BalanceLineAsset"
								},
								"examples": {
									"Existing Stellar balance line": {
										"value": {
											"balance": "10.5",
											"limit": "1000000",
											"asset_code": "USDC",
											"asset_issuer": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF",
											"buying_liabilities": "0.0",
											"selling_liabilities": "0.0",
											"last_modified_ledger": 123456,
											"is_authorized": true,
											"is_authorized_to_maintain_liabilities": true,
											"is_clawback_enabled": false
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens",
					"Transfers"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/check/algorand/optin": {
			"get": {
				"operationId": "CheckAlgorandAssetOptIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								},
								"examples": {
									"Algorand opt-in check result": {
										"value": true
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HttpExceptionBody"
								}
							}
						}
					}
				},
				"tags": [
					"Tokens",
					"Transfers"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sender",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"default": "asset",
							"type": "string",
							"enum": [
								"asset",
								"app"
							]
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}